+++ djmoney/contrib/exchange/models.py
@@ -1,10 +1,11 @@
 from django.conf import settings
+from django.core.cache import cache
 from django.core.exceptions import ImproperlyConfigured
 from django.db import models
 from django.utils.module_loading import import_string
 
 from djmoney._compat import text_type
-from djmoney.settings import EXCHANGE_BACKEND
+from djmoney.settings import EXCHANGE_BACKEND, RATES_CACHE_TIMEOUT
 
 from .exceptions import MissingRate
 
@@ -51,6 +52,16 @@ def get_rate(source, target, backend=None):
     """
     if backend is None:
         backend = get_default_backend_name()
+    key = 'djmoney:get_rate:%s:%s:%s' % (source, target, backend)
+    result = cache.get(key)
+    if result is not None:
+        return result
+    result = _get_rate(source, target, backend)
+    cache.set(key, result, RATES_CACHE_TIMEOUT)
+    return result
+
+
+def _get_rate(source, target, backend):
     if text_type(source) == text_type(target):
         return 1
     try: