+++ djmoney/models/fields.py
@@ -112,6 +112,9 @@ class MoneyFieldProxy(object):
         return data[self.field.name]
 
     def __set__(self, obj, value):  # noqa
+        if value is not None and self.field._currency_field.null and not isinstance(value, MONEY_CLASSES):
+            # For nullable fields we need either both NULL amount and currency or both NOT NULL
+            raise ValueError('Missing currency value')
         if isinstance(value, BaseExpression):
             if isinstance(value, Value):
                 value = self.prepare_value(obj, value.value)
@@ -174,7 +177,7 @@ class MoneyField(models.DecimalField):
                  currency_field_name=None, **kwargs):
         nullable = kwargs.get('null', False)
         default = self.setup_default(default, default_currency, nullable)
-        if not default_currency:
+        if not default_currency and default is not None:
             default_currency = default.currency
 
         self.default_currency = default_currency
@@ -255,7 +258,7 @@ class MoneyField(models.DecimalField):
         currency_field = CurrencyField(
             price_field=self,
             default=self.default_currency, editable=False,
-            choices=self.currency_choices
+            choices=self.currency_choices, null=self.default_currency is None
         )
         currency_field.creation_counter = self.creation_counter - 1
         currency_field_name = get_currency_field_name(name, self)