+++ djmoney/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '0.14'
+__version__ = '0.14.1'
 default_app_config = 'djmoney.apps.MoneyConfig'