+++ docs/changes.rst
@@ -6,18 +6,27 @@ Changelog
 `0.14`_ - unreleased
 --------------------
 
+Added
+~~~~~
+
+- Caching of exchange rates. `#398`_ (`Stranger6667`_)
+
+Changed
+~~~~~~~
+
+- Added support for nullable ``CurrencyField``. `#260`_ (`Stranger6667`_)
+
 Fixed
 ~~~~~
 
 - Same currency conversion getting MissingRate exception `#418`_ (`humrochagf`_)
-- Added support for nullable ``CurrencyField``. `#260`_ (`Stranger6667`_)
+- `TypeError` during templatetag usage inside a for loop on Django 2.0. `#402`_ (`f213`_)
 
 Removed
 ~~~~~~~
 
 - Support for Python 3.3 `#410`_ (`benjaoming`_)
 - Deprecated ``choices`` argument from ``djmoney.forms.fields.MoneyField``. Use ``currency_choices`` instead. (`Stranger6667`_)
-
 
 `0.13.5`_ - 2018-05-19
 ----------------------
@@ -551,8 +560,10 @@ Added
 
 .. _#412: https://github.com/django-money/django-money/issues/412
 .. _#410: https://github.com/django-money/django-money/issues/410
+.. _#402: https://github.com/django-money/django-money/issues/402
 .. _#400: https://github.com/django-money/django-money/issues/400
 .. _#399: https://github.com/django-money/django-money/issues/399
+.. _#398: https://github.com/django-money/django-money/issues/398
 .. _#392: https://github.com/django-money/django-money/issues/392
 .. _#388: https://github.com/django-money/django-money/issues/388
 .. _#385: https://github.com/django-money/django-money/issues/385
@@ -645,6 +656,7 @@ Added
 .. _elcolie: https://github.com/elcolie
 .. _eriktelepovsky: https://github.com/eriktelepovsky
 .. _evenicoulddoit: https://github.com/evenicoulddoit
+.. _f213: https://github.com/f213
 .. _glarrain: https://github.com/glarrain
 .. _graik: https://github.com/graik
 .. _gonzalobf: https://github.com/gonzalobf